transfer get Geolocation route and controller to routes js and controllers js
