+++ backend/routes/routes.js
@@ -1,8 +1,9 @@
 const express = require("express");
-const { addCoordinate, getCoordinates } = require("../controller/controllers");
+const { getGeoLocation , addCoordinate, getCoordinates } = require("../controller/controllers");
 const router = express.Router();
 
 router.post("/addCoordinate", addCoordinate);
 router.get("/getCoordinates", getCoordinates);
+router.get('/map', getGeoLocation);
 
 module.exports = router;
